[NTOSKRNL] Clean up BACKTRACE_INFO definition
This type was needlessly placed in ntoskrnl.h, despite its only
being used in ke\bug.c. This caused a rebuild of the entire kernel
if the definition of this type changed. I have moved it into ke\bug.c
for this reason; I have also reduced the number of casts required in
KiDoStackBacktrace.